+++ Editor/d4rkAvatarOptimizer.cs
@@ -1749,6 +1749,15 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                 continue;
             }
             var layer = fxLayerLayers[i];
+            if (layer.syncedLayerIndex != -1)
+            {
+                errorMessages[i].Add($"synced with layer {layer.syncedLayerIndex}");
+                if (layer.syncedLayerIndex >= 0 && layer.syncedLayerIndex < fxLayerLayers.Length)
+                {
+                    errorMessages[layer.syncedLayerIndex].Insert(0, $"layer {i} is synced with this layer");
+                }
+                continue;
+            }
             var stateMachine = layer.stateMachine;
             if (stateMachine == null)
             {
@@ -2098,6 +2107,8 @@ public class d4rkAvatarOptimizer : MonoBehaviour
             if (i <= 2 && MMDCompatibility)
                 break;
             var layer = fxLayerLayers[i];
+            if (layer.syncedLayerIndex != -1)
+                continue;
             bool isNotFirstLayerOrLastNonUselessLayerCanBeFirst = i != 0 ||
                 (lastNonUselessLayer < fxLayerLayers.Length && fxLayerLayers[lastNonUselessLayer].avatarMask == layer.avatarMask
                     && fxLayerLayers[lastNonUselessLayer].defaultWeight == 1 && !isAffectedByLayerWeightControl.Contains(lastNonUselessLayer));
@@ -2129,6 +2140,13 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                 continue;
             }
             lastNonUselessLayer = i;
+        }
+        for (int i = 0; i < fxLayerLayers.Length; i++)
+        {
+            if (fxLayerLayers[i].syncedLayerIndex != -1)
+            {
+                uselessLayers.Remove(fxLayerLayers[i].syncedLayerIndex);
+            }
         }
         Profiler.EndSection();
         return cache_FindUselessFXLayers = uselessLayers;